pyproject.toml
[project]
name = "wyoming-parakeet"
version = "2.5.0"
description = "Wyoming Server for Parakeet"
readme = "README.md"
requires-python = ">=3.8.1,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Michael Hansen", email = "mike@rhasspy.org"}
]
keywords = ["rhasspy", "wyoming", "parakeet", "stt"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Text Processing :: Linguistic",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "wyoming>=1.5.3",
    "nemo-toolkit[asr]>=1.18.0",
    "torch>=1.13.0",
    "torchaudio>=0.13.0",
    "librosa>=0.9.0"
]

[project.urls]
Homepage = "http://github.com/rhasspy/wyoming-parakeet"

[tool.setuptools]
platforms = ["any"]
zip-safe  = true
include-package-data = true

[tool.setuptools.packages.find]
include = ["wyoming_parakeet"]
exclude = ["tests"]

[project.scripts]
wyoming-parakeet = "wyoming_parakeet.__main__:run"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
check_untyped_defs = true
disallow_untyped_defs = true

[project.optional-dependencies]
dev = [
    "black==22.12.0",
    "flake8==6.0.0",
    "isort==5.11.3",
    "mypy==0.991",
    "pylint==2.15.9",
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
]
transformers = [
    "transformers[torch]==4.52.4",
]
